+++ constants/index.js
@@ -1,8 +1,13 @@
+const POOLS = require('./pools.json');
+
 const THIRTY_DAYS = 2592000;
 
 const DISTRIBUTOR = '0xE527972DD80970bb90977879658FB83aD568Bb28';
+const REWARD_TOKEN = '0x0BD802635eb9cEB3fCBe60470D2857B86841aab6';
 
 module.exports = {
   DISTRIBUTOR,
+  POOLS,
+  REWARD_TOKEN,
   THIRTY_DAYS,
 };
