--- scripts/refreshAllPools/pools.json
@@ -1,18 +1,0 @@
-[
-  {
-    "pool": "veusd-vet",
-    "address": "0xf1be58861b4bcacd6c7d026ba3de994361f5d3aa",
-    "monthlyRate": 5000,
-    "rewardTokens": [
-      "0x0BD802635eb9cEB3fCBe60470D2857B86841aab6"
-    ]
-  },
-  {
-    "pool": "veusd-vet",
-    "address": "0xf1be58861b4bcacd6c7d026ba3de994361f5d3aa",
-    "monthlyRate": 5000,
-    "rewardTokens": [
-      "0x0BD802635eb9cEB3fCBe60470D2857B86841aab6"
-    ]
-  }
-]
